fix: address PR review issues (layer violation, readonly, ProjectInfo.path, deleteSession void)
