+++ src/features/agent/reactions.test.ts
@@ -3,7 +3,8 @@ import type {Logger} from '../../shared/logger.js'
 import type {ReactionContext} from './types.js'
 
 import {afterEach, beforeEach, describe, expect, it, vi} from 'vitest'
-import {createMockLogger, createMockOctokit} from '../../shared/test-helpers.js'
+import {createMockOctokit} from '../../services/github/test-helpers.js'
+import {createMockLogger} from '../../shared/test-helpers.js'
 import {
   acknowledgeReceipt,
   addEyesReaction,
