+++ src/features/delegated/branch.test.ts
@@ -1,6 +1,7 @@
 import type {Logger} from '../../shared/logger.js'
 import {beforeEach, describe, expect, it, vi} from 'vitest'
-import {createMockLogger, createMockOctokit} from '../../shared/test-helpers.js'
+import {createMockOctokit} from '../../services/github/test-helpers.js'
+import {createMockLogger} from '../../shared/test-helpers.js'
 import {branchExists, createBranch, deleteBranch, generateBranchName} from './branch.js'
 
 describe('createBranch', () => {
