+++ src/features/triggers/issue-comment.test.ts
@@ -3,7 +3,8 @@ import type {Logger} from '../../shared/logger.js'
 import type {TriggerContext} from './types.js'
 import {beforeEach, describe, expect, it, vi} from 'vitest'
 import {normalizeEvent} from '../../services/github/context.js'
-import {createMockLogger, createMockOctokit} from '../../shared/test-helpers.js'
+import {createMockOctokit} from '../../services/github/test-helpers.js'
+import {createMockLogger} from '../../shared/test-helpers.js'
 import {createIssueCommentCreatedEvent} from './__fixtures__/payloads.js'
 import {handleIssueComment} from './issue-comment.js'
