+++ src/services/setup/gh-auth.test.ts
@@ -1,6 +1,7 @@
 import type {ExecAdapter, Logger} from './types.js'
 import {afterEach, beforeEach, describe, expect, it, vi} from 'vitest'
-import {createMockLogger, createMockOctokit} from '../../shared/test-helpers.js'
+import {createMockLogger} from '../../shared/test-helpers.js'
+import {createMockOctokit} from '../github/test-helpers.js'
 import {configureGhAuth, configureGitIdentity, getBotLogin, getBotUserId} from './gh-auth.js'
 
 function createMockExecAdapter(overrides: Partial<ExecAdapter> = {}): ExecAdapter {
